refactor: Rename classes and update import paths; remove unused ongoingEvents widget
